Define matcher macros in Matchers::Macros module.
