+++ lib/rspec/sleeping_king_studios/matchers/built_in/respond_to.rb
@@ -1,8 +1,9 @@
 # lib/rspec/sleeping_king_studios/matchers/built_in/respond_to.rb
 
 require 'rspec/sleeping_king_studios/matchers/built_in/respond_to_matcher'
+require 'rspec/sleeping_king_studios/matchers/macros'
 
-module RSpec::SleepingKingStudios::Matchers
+module RSpec::SleepingKingStudios::Matchers::Macros
   # @see RSpec::SleepingKingStudios::Matchers::BuiltIn::RespondToMatcher#matches?
   def respond_to *expected
     RSpec::SleepingKingStudios::Matchers::BuiltIn::RespondToMatcher.new *expected
