+++ lib/rspec/sleeping_king_studios/matchers/core/alias_method.rb
@@ -1,8 +1,9 @@
 # lib/rspec/sleeping_king_studios/matchers/core/alias_method.rb
 
 require 'rspec/sleeping_king_studios/matchers/core/alias_method_matcher'
+require 'rspec/sleeping_king_studios/matchers/macros'
 
-module RSpec::SleepingKingStudios::Matchers
+module RSpec::SleepingKingStudios::Matchers::Macros
   # @see RSpec::SleepingKingStudios::Matchers::Core::AliasMethodMatcher#matches?
   def alias_method expected
     RSpec::SleepingKingStudios::Matchers::Core::AliasMethodMatcher.new expected
