+++ lib/rspec/sleeping_king_studios/matchers/core/have_predicate.rb
@@ -1,8 +1,9 @@
 # lib/rspec/sleeping_king_studios/matchers/core/have_predicate.rb
 
 require 'rspec/sleeping_king_studios/matchers/core/have_predicate_matcher'
+require 'rspec/sleeping_king_studios/matchers/macros'
 
-module RSpec::SleepingKingStudios::Matchers
+module RSpec::SleepingKingStudios::Matchers::Macros
   # @see RSpec::SleepingKingStudios::Matchers::Core::HavePredicateMatcher#matches?
   def have_predicate expected
     RSpec::SleepingKingStudios::Matchers::Core::HavePredicateMatcher.new expected
