+++ lib/rspec/sleeping_king_studios/matchers/core/have_reader.rb
@@ -1,8 +1,9 @@
 # lib/rspec/sleeping_king_studios/matchers/core/have_reader.rb
 
 require 'rspec/sleeping_king_studios/matchers/core/have_reader_matcher'
+require 'rspec/sleeping_king_studios/matchers/macros'
 
-module RSpec::SleepingKingStudios::Matchers
+module RSpec::SleepingKingStudios::Matchers::Macros
   # @see RSpec::SleepingKingStudios::Matchers::Core::HaveReaderMatcher#matches?
   def have_reader expected
     RSpec::SleepingKingStudios::Matchers::Core::HaveReaderMatcher.new expected
