+++ spec/rspec/sleeping_king_studios/matchers/macros/be_kind_of_spec.rb
@@ -5,7 +5,7 @@ require 'spec_helper'
 require 'rspec/sleeping_king_studios/matchers/built_in/be_kind_of'
 require 'rspec/sleeping_king_studios/matchers/core/alias_method'
 
-RSpec.describe RSpec::SleepingKingStudios::Matchers do
+RSpec.describe RSpec::SleepingKingStudios::Matchers::Macros do
   let(:matcher_class) do
     RSpec::SleepingKingStudios::Matchers::BuiltIn::BeAKindOfMatcher
   end # let
