+++ spec/rspec/sleeping_king_studios/matchers/macros/include_spec.rb
@@ -5,7 +5,7 @@ require 'rspec/sleeping_king_studios/matchers/built_in/respond_to'
 
 require 'rspec/sleeping_king_studios/matchers/built_in/include'
 
-RSpec.describe RSpec::SleepingKingStudios::Matchers do
+RSpec.describe RSpec::SleepingKingStudios::Matchers::Macros do
   let(:matcher_class) do
     ::RSpec::SleepingKingStudios::Matchers::BuiltIn::IncludeMatcher
   end # let
