+++ spec/rspec/sleeping_king_studios/matchers/macros/respond_to_spec.rb
@@ -4,7 +4,7 @@ require 'spec_helper'
 
 require 'rspec/sleeping_king_studios/matchers/built_in/respond_to'
 
-RSpec.describe RSpec::SleepingKingStudios::Matchers do
+RSpec.describe RSpec::SleepingKingStudios::Matchers::Macros do
   let(:matcher_class) do
     ::RSpec::SleepingKingStudios::Matchers::BuiltIn::RespondToMatcher
   end # let
